+++ src/cmds/say.js
@@ -1,8 +1,8 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
-const { EmbedBuilder, WebhookClient } = require('discord.js');
+const { EmbedBuilder } = require('discord.js');
 const config = require('../../config');
 
-const {cmdHook} = require("../hooks");
+const { cmdHook } = require('../utils/WebhookManager');
 
 const sayData = new SlashCommandBuilder()
   .setName("say")
@@ -17,14 +17,14 @@ module.exports = {
   data: sayData,
   
   async execute(interaction) {
-    const content = interaction.options.getString("訊息內容"); 
+    const content = interaction.options.getString("訊息內容");
 
     if(content === '@everyone'){
       await interaction.reply({ content: '還想 @everone 啊，你是吃了熊心豹子膽嗎？', ephemeral: true });
       return;
     }
     
-    await interaction.reply({ content: "訊息已傳送", ephemeral: true });
+    await interaction.reply({ content: '訊息傳出去了喔：）', ephemeral: true });
     await interaction.channel.send({ content: content, allowedMentions: { parse: [] } });
 
     const cmdHookEmbed = new EmbedBuilder()
