+++ src/cmds/ticTacToe.js
@@ -1,37 +1,36 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
-const { EmbedBuilder, WebhookClient } = require('discord.js')
+const { EmbedBuilder } = require('discord.js')
 const { DjsTicTacToe } = require('@hizollo/games');
 const { ticTacToe } = require('../data/GameStrings.json');
-const config = require('../../config');
 
-const {cmdHook} = require("../hooks");
+const { cmdHook } = require('../utils/WebhookManager');
 
 const ticTacToeData = new SlashCommandBuilder()
   .setName("tictactoe")
   .setDescription("遊玩一場圈圈叉叉遊戲！")
   .addIntegerOption(option => 
-      option.setName("盤面大小")
-      .setDescription("請輸入盤面大小 (1~4)")
-      .setRequired(true)
+    option.setName("盤面大小")
+    .setDescription("請輸入盤面大小 (1~4)")
+    .setRequired(true)
   )
   .addUserOption(option => 
-      option.setName("對手1")
-      .setDescription("對手 1")
-      .setRequired(true)
+    option.setName("對手1")
+    .setDescription("對手 1")
+    .setRequired(true)
   )
   .addUserOption(option =>
-      option.setName("對手2")
-      .setDescription("對手 2")
-      .setRequired(false)
-  );
+    option.setName("對手2")
+    .setDescription("對手 2")
+    .setRequired(false)
+  )
 
 module.exports = {
   data: ticTacToeData,
 
   async execute(interaction){
     const user = interaction.user;
-    const p2 = interaction.options.getUser("對手 1");
-    const p3 = interaction.options.getUser("對手 2");
+    const p2 = interaction.options.getUser("對手1");
+    const p3 = interaction.options.getUser("對手2");
     const boardSize = interaction.options.getInteger("盤面大小");
 
     async function duplicateIds(){
@@ -55,7 +54,7 @@ module.exports = {
        symbol: "⭕",
     }];
 
-    if(user.id === p2.id || user === p3?.id){
+    if([p2.id, p3?.id ?? ''].includes(user.id)){
       duplicateIds();
     }
 
@@ -83,11 +82,11 @@ module.exports = {
       .setColor(0x00bfff)
       .setDescription("Command: `/tictactoe`")
       .addFields(
-          { name: "User Tag", value: interaction.user.tag },
-          { name: "User ID", value: interaction.user.id },
-          { name: "Guild Name", value: interaction.guild.name },
-          { name: "Guild ID", value: interaction.guild.id },
-          { name: "Players", value: `${user.tag} & ${p2.tag}`}
+        { name: "User Tag", value: interaction.user.tag },
+        { name: "User ID", value: interaction.user.id },
+        { name: "Guild Name", value: interaction.guild.name },
+        { name: "Guild ID", value: interaction.guild.id },
+        { name: "Players", value: `${user.tag} & ${p2.tag} & ${p3?.tag}`}
       )
       .setTimestamp()
       .setFooter({ text: 'Shard#1' });
